% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-convert-utm-to-latlon.R
\name{convert_utm_to_latlon}
\alias{convert_utm_to_latlon}
\title{Convert UTM coordinates to latitude/longitude}
\usage{
convert_utm_to_latlon(site_info)
}
\arguments{
\item{site_info}{Data table with UTM coordinates and zone.}
}
\value{
Data table with added LATITUDE and LONGITUDE columns.
}
\description{
Converts UTM easting, northing, and zone columns in site_info to latitude and longitude.
}
