% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{map_layer}
\alias{map_layer}
\title{Mapping table for INPUT_LAYER variables}
\format{
A data.table with the same structure as \code{map_polygon}.
}
\source{
inst/extdata/nfitovdyp_mapping_layer.csv
}
\usage{
map_layer
}
\description{
This dataset contains the variable mapping between National Forestry Inventory (NFI) data
and VDYP INPUT_LAYER variables. It specifies how to calculate or extract species composition,
height, basal area, and other tree-level attributes for VDYP input.
}
\keyword{datasets}
