% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{map_poly}
\alias{map_poly}
\title{Mapping table for INPUT_POLY variables}
\format{
A data.table with columns:
\describe{
\item{VDYP Variable}{The target VDYP variable name}
\item{NFI Variable (maps to)}{NFI source variable(s), separated by ";" if multiple}
\item{NFI csv file}{The NFI CSV file(s) to read}
\item{Type}{Expected data type (integer, numeric, character, etc.)}
\item{Defaults To}{Default value if no data is available}
\item{Description}{Description or notes (not used in calculation)}
}
}
\source{
inst/extdata/nfitovdyp_mapping_polygon.csv
}
\usage{
map_poly
}
\description{
This dataset contains the variable mapping between National Forestry Inventory (NFI) data
and VDYP INPUT_POLYGON variables. It includes the VDYP variable name, NFI source variables,
required data files, variable type, default values, and a short description of the procedure.
}
\keyword{datasets}
