% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfi-to-vdyp-data-prep-new.R
\name{nfi_to_vdyp}
\alias{nfi_to_vdyp}
\title{Convert NFI data to VDYP input format}
\usage{
nfi_to_vdyp(
  nfi_folder,
  output_path = file.path(getwd(), "nfi_to_vdyp_data"),
  remeasurement_number = NULL
)
}
\arguments{
\item{nfi_folder}{Path to folder containing NFI CSV files.}

\item{output_path}{Path to folder where VDYP input CSVs will be written. Default is current working directory + "nfi_to_vdyp_data".}

\item{remeasurement_number}{(Optional) measurement number to filter the data.}
}
\value{
No return value. Writes two CSV files to disk.
}
\description{
This function reads National Forest Inventory (NFI) ground plot data,
transforms it into the input format required by \href{https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/growth-and-yield-modelling/variable-density-yield-projection-vdyp/publications-and-support}{Variable Density Yield Projection (VDYP 7)},
and writes INPUT_POLY.csv and INPUT_LAYER.csv to an output directory.
}
\details{
The function uses internal mapping tables (\code{map_polygon} and \code{map_layer}) to translate
National Forestry Inventory (NFI) data into the format required by VDYP.

Key handling rules include:
\itemize{
\item \strong{Direct mappings:} For most variables, NFI values are used as-is or default to the specified fallback if missing.
\item \strong{Calculated fields:} Some variables require computation, e.g., percent standing dead trees is calculated as \code{plotvol_standdead / (plotvol_standlive + plotvol_standdead) * 100}.
\item \strong{UTM conversions:} LATITUDE and LONGITUDE are computed from UTM coordinates.
\item \strong{BEC zones:} BEC_ZONE_CODE is assigned using a GIS intersect between NFI plot coordinates and BEC zone polygons.
\item \strong{Species composition:} For species codes and percentages, the function harmonizes small and large tree datasets, recalculates basal area per hectare for dbh >= 7.5cm (and dbh >= 12.5cm), computes relative percentages, and ranks species dominance.

Important: Recalculation of basal area is done separately for large tree and small trees because large trees are measure on the large tree plot which is 11.28m in diameter (or 0.04 ha), while small trees are measured on the small tree plot which is half the diameter (5.64 m) and a quarter of the area (0.01 ha)
\item \strong{Height and volume metrics:} Metrics like LOREY_HEIGHT_75, BASAL_AREA_125, and WS_VOL_PER_HA_75/125 are computed using filtered live tree data.
\item \strong{Post-processing or fillers:} Some fields are populated with placeholders (e.g., PLOT_NUMBER or FEATURE_ID) for downstream processing or tracing.
}

The full description of each mapping rule is stored in the \code{Description} column of
the \code{map_polygon} and \code{map_layer} datasets included in the package.
}
\examples{
\dontrun{
nfi_to_vdyp("path/to/nfi_data")
}
}
